requirements.txt
Django==5.1.3
djangorestframework==3.15.2
drf-yasg==1.21.8
mysqlclient==2.2.6
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1